Use original __info__ and module_info functions in patched modules, remove the Kernel patch (#87)
Fixes EXW-237

So far, in the patched modules, we used `__info__/1`, `module_info/0`
and `module_info/1` from the patch, so that
`PatchedModule.module_info(:functions)` would only return the patched
functions. That's why macros weren't working in patched modules unless
they're patched too. This PR fixes it by using the original versions of
`__info__` and friends. It's not perfect, as if we added a
macro/function in a patch that doesn't exist in the original module, it
won't be present in `__info__` and friends, but I don't think it's the
case, at least for now. The proper fix would require generating
`__info__` and friends based on the final, patched module, which is not
trivial.