DEMO - CI for building and deploying docker image (#99)
This PR brings working workflows for building and deploying Docker
images containing Nginx serving html files for demo.

Staging (single VM):
[popcorn.swmtest.xyz](https://popcorn.swmtest.xyz)
(currently deployed on `push` for the `deploy` branch - modify it to
your preference)

Production (3 VMs in different datacenters in a round-robin DNS
resolution):
[popcorn.swmansion.com](https://popcorn.swmansion.com)
(currently deployed on semver tags `v.*.*.*`)

The branch contains built HTML files that I received zipped, and to my
knowledge, they _do not work as expected_, i.e., repl isn't interpreting
Elixir code, but it is hanging on _evaluating_.
Anyway, these files should be replaced by a job stage in the CI that
generates all the HTML files and passes them to the Docker image build
phase - to my best knowledge, this was about to be @mat-hek task.

Potential improvements:
- both workflows currently build and deploy `latest` image tag - this
can be problematic in some cases
- production environment _will have problems_ with the LE certificates
generation due to the round-robin (but not sooner than on 02.07.2025)

---------

Co-authored-by: Jakub Gonet <jakub.gonet@swmansion.com>